Took out the mocking component of the project.
Got rid of everywhere I was using mock data. Now, should be able to just
modify the ProjectLocation and ProjectUpdateLocation variables, and it
should be ready to use. Still have questions about getting feedback from
user about which dependency to update.

Have not retested it yet, so it might not work. Closes #41, Closes #42,
Closes #45, Closes #48.